bot_task/tf2_taunttask: expand task
Bots will try to stick to the player when doing dance taunts now. There was also an attempt to get them to stop taunting after a short while, this is somewhat broken in its current state though. Ideally we should be able to do whatever the removecond console command does, as that stops the taunt without issues